ci/calculate-version-script/requirements.txt
toml
pyyaml
towncrier
